telco-churn-production/requirements.txt
# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Machine Learning
xgboost>=2.0.0
catboost>=1.2.0
imbalanced-learn>=0.11.0

# Experiment Tracking
mlflow>=2.9.0

# Spark
pyspark>=3.5.0
pyarrow>=14.0.0

# Pipeline & Utilities
joblib>=1.3.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0